feat(tests): host-side unit test scaffold + CI workflow
Adds a Unity-based unit test suite that runs on a developer machine /
in CI without any toolchain, hardware, or flash time. Mirrors the
pattern already in use in isc-fs/IFS08-CE-AMS so both repos share the
same testing conventions.

Suite covers pure-logic invariants we historically learn about only
when a carrier bricks on the bench:

- bl_isotp:  SF / FF / CF framing, sequence-number validation,
             reassembly overflow, TX segmentation iterator.
- bl_fwinfo: magic + record_version acceptance / rejection, forward-
             compat policy on future minor revisions.
- bl_nvm:    dedup, tombstones, live_count tracking, value-length
             rejection, plus end-to-end coverage of fix #13's new
             bl_nvm_compact_replace_meta() and bl_nvm_format(), plus
             the bl_nvm_write retry-via-compaction defense added on
             top of it. The HAL stub enforces the STM32 1→0-only
             programming rule, so the metadata-rewrite regression
             from issue #52 is locked down by an explicit test.

Bug surfaced + tracked, not fixed in this PR:

- bl_isotp_rx_feed() never sets rx->deadline_ms when FF lands, so
  bl_isotp_rx_tick() reports timeout immediately. The corresponding
  test is `TEST_IGNORE`'d with a pointer to issue #54 (separate PR
  for the fix). The infra PR stays green.

Three minor portability tweaks in production code to enable the host
build (each a no-op on STM32 where sizeof(uintptr_t) == sizeof(uint32_t)):

- bl_nvm.c::program_flashword: addr parameter widened to uintptr_t.
- bl_nvm.c::program_flashword: data cast widened to (uintptr_t).
- bl_flash.c: two `(uint32_t)ptr` casts widened to `(uintptr_t)ptr`
  in bl_flash_write + bl_flash_write_metadata.

These match what we'd want long-term anyway for any 64-bit-friendly
flash-mock harness; they don't change generated code on Cortex-M.

CI workflow at .github/workflows/build-and-test.yml:

- firmware-build (×2): arm-none-eabi-gcc 14.3 + Ninja, Debug + Release.
  Catches warnings-as-errors, link failures, memory-budget regressions.
  Uploads .elf / .bin / .hex as artifacts (14-day retention).
- host-tests: Unity + ctest. Must be green for any PR to merge once
  branch protection is configured.

Note: org GitHub Actions billing is currently paused. The workflow
file lands so it activates the moment billing resumes — no further
config needed at that point.

Run locally:
    cmake -B build-tests -S tests/unit
    cmake --build build-tests
    ctest --test-dir build-tests --output-on-failure
    # → 25 passed / 1 ignored / 0 failed in 0.34 s

+++ Core/Src/bl_flash.c
@@ -142,9 +142,11 @@ bl_flash_status_t bl_flash_write(uint32_t addr,
             memset(fw + chunk, 0xFFU, BL_FLASHWORD_BYTES - chunk);
         }
 
+        /* `(uintptr_t)` rather than `(uint32_t)`: identity on STM32,
+         * lossless on a 64-bit host (unit tests). */
         HAL_StatusTypeDef st = HAL_FLASH_Program(FLASH_TYPEPROGRAM_FLASHWORD,
                                                  addr + offset,
-                                                 (uint32_t)fw);
+                                                 (uintptr_t)fw);
         if (st != HAL_OK) {
             HAL_FLASH_Lock();
             return BL_FLASH_ERR_HARDWARE;
@@ -217,9 +219,11 @@ bl_flash_status_t bl_flash_write_metadata(uint32_t size,
         return BL_FLASH_ERR_HARDWARE;
     }
 
+    /* `(uintptr_t)` rather than `(uint32_t)`: identity on STM32,
+     * lossless on a 64-bit host (unit tests). */
     HAL_StatusTypeDef st = HAL_FLASH_Program(FLASH_TYPEPROGRAM_FLASHWORD,
                                              BL_APP_METADATA_ADDR,
-                                             (uint32_t)meta);
+                                             (uintptr_t)meta);
     HAL_FLASH_Lock();
 
     return (st == HAL_OK) ? BL_FLASH_OK : BL_FLASH_ERR_HARDWARE;
